Remove logic that rejects connections on limit reached since may cause issues when running standalone
Since server may be closing a connection when client closes a connection and re opens another it may happen that the number of connections open for just one client may be greater than 1, so this causes for example blaze meter tests to fail. To avoid this we just remove the limitation on number of client connections to accept and only leave the limitation of the number of concurrent connections to process at a given time.