Fix logic to disconnect RTE connections only once per thread group iteration
iterationStart method is invoked once for each sampler in each controller (thread group, loop controller, if controller, etc) start. To avoid disconnecting RTE when for example an RTE sampler is included in an IF controller, and to avoid unnecessary
calls, we only reset connection when the sampler is the first sampler in the thread group iteration.