Request headers, responseBody, requestBody fixed
At: RTESampler class, there is an Override called sample. I refactor that class to use RTESamplerResult.class but is not working as it was expected.(It does not show sampleResult, screen, responseBody and neither responseHeaders ).
On the other side. RTERecorder it might be ready to link up with the Terminal Gui. It shows sampleResult as we expect, adding samplers into RecordingController and showing the results in ViewResultsTree, all of it with dummy values for now.