Allow to connect without waiting for keyboard to be unlocked
This is required to be able to allow users to specify other conditions if needed in some test plan, and additionally is useful when recording to provide feedback ASAP to user when connected to server and to avoid locks issues with tn5250 connection getting stuck (due to swing dispatch thread being used for connecting).

Additionally, fix removal of listeners for tn3270 and avoid memory leak, provide real time notification to emulator when an exception arises, fix integration tests intermittently failing on setting up jmeter testing environment.