Add new test element for recording rte test plans
A new major feature has been added that will greatly simplify creating RTE test plans: A RECORDER ELEMENT! This recorder allows you to interact with a mainframe application (currently supporting tn3270 and tn5250 protocols) as you usually do with a terminal emulator and record every interaction with the mainframe application as a test plan, that can later on be used for functional or load testing. Please check the updated readme for a guide on how to use it and additional details.

In addition to the recorder, following fixes have been included in existing sampler and protocol support:
- Fix tn3270 protocol support:
  - Fix handling alternate screen dimensions
  - Fix handling enter attention keys when no field is modified
  - Fix usage of logging framework to avoid unintentional logging to stdout and provide better control on logged information.
  - Fix returned fields to avoid potential wrapping (eg: field starting at row 22 and ending at row 2).
  - Fix enhanced protocol headers sequence numbering for some command replies.
- Fix cursor positioning when a field text is filled to match emulator behavior
- Fix connection time to be more precise and don’t count sync time (time until keyboard is unlocked)
- Fixed failing connection to a mainframe application that does not unlock keyboard (may be only requesting some attention key)
- Performed improvements to simplify the code.
- Updated all dependencies of the plugin