RTE-60 adding logic to search for a close field
current logic is looking for closer field on the right when the offset value is positive (meaning that we are going to skip forward a field). In the other hand when offset value is negative the logic will search for the closer field on the left. In remote case, offset has value of zero, then it will search for closer field on the right.